Load variables used to render autoyast in separate class.
